@@ -46,18 +46,15 @@ else()
 endif()
 
 # C/C++
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
+set(CMAKE_EXPERIMENTAL_CXX_MODULE_CMAKE_API "2182bf5c-ef0d-489a-91da-49dbc3090d2a")
+set(CMAKE_EXPERIMENTAL_CXX_MODULE_DYNDEP 1)
 add_definitions(-DREDUMPER_VERSION_MAJOR=${PROJECT_VERSION_MAJOR}
                 -DREDUMPER_VERSION_MINOR=${PROJECT_VERSION_MINOR}
                 -DREDUMPER_VERSION_PATCH=${PROJECT_VERSION_PATCH}
                 -DREDUMPER_VERSION_BUILD=${REDUMPER_VERSION_BUILD})
-
-# fmt
-# remove this after gcc/clang gets full std::format support
-set(FMT_INCLUDE "${CMAKE_CURRENT_SOURCE_DIR}/fmt/include")
-add_definitions(-DFMT_HEADER_ONLY)
 
 # Visual Studio settings
 if(MSVC)
@@ -86,7 +83,6 @@ endif()
 # generated source file
 set(DRIVEOFFSETS_FN "${CMAKE_CURRENT_BINARY_DIR}/driveoffsets.inc")
 add_executable(generate_offsets generate_offsets.cc)
-target_include_directories(generate_offsets PUBLIC ${FMT_INCLUDE})
 add_custom_command(
 	OUTPUT ${DRIVEOFFSETS_FN}
 	COMMAND generate_offsets "${CMAKE_CURRENT_SOURCE_DIR}/driveoffsets.txt"
@@ -174,7 +170,6 @@ set(sources
 set(includes
 	${CMAKE_CURRENT_SOURCE_DIR}
 	${CMAKE_CURRENT_BINARY_DIR}
-	${FMT_INCLUDE}
 )
 
 add_executable(redumper ${sources})
