@@ -342,17 +342,17 @@ void write_tracks(std::vector<TrackEntry> &track_entries, TOC &toc, std::fstream
 
 			// add session number to lead-in/lead-out track string to make filename unique
 			if(lilo && toc.sessions.size() > 1)
-				track_string = fmt::format("{}.{}", track_string, s.session_number);
+				track_string = std::format("{}.{}", track_string, s.session_number);
 
-			std::string track_name = fmt::format("{}{}.bin", options.image_name, toc.TracksCount() > 1 || lilo ? fmt::format(" (Track {})", track_string) : "");
+			std::string track_name = std::format("{}{}.bin", options.image_name, toc.TracksCount() > 1 || lilo ? std::format(" (Track {})", track_string) : "");
 			LOG("writing \"{}\"", track_name);
 
 			if(std::filesystem::exists(std::filesystem::path(options.image_path) / track_name) && !options.overwrite)
-				throw_line(fmt::format("file already exists ({})", track_name));
+				throw_line(std::format("file already exists ({})", track_name));
 
 			std::fstream fs_bin(std::filesystem::path(options.image_path) / track_name, std::fstream::out | std::fstream::binary);
 			if(!fs_bin.is_open())
-				throw_line(fmt::format("unable to create file ({})", track_name));
+				throw_line(std::format("unable to create file ({})", track_name));
 
 			TrackEntry track_entry;
 			track_entry.filename = track_name;
@@ -437,7 +437,7 @@ void write_tracks(std::vector<TrackEntry> &track_entries, TOC &toc, std::fstream
 
 				fs_bin.write((char *)sector.data(), sector.size());
 				if(fs_bin.fail())
-					throw_line(fmt::format("write failed ({})", track_name));
+					throw_line(std::format("write failed ({})", track_name));
 			}
 
 			for(auto const &d : descramble_errors)
@@ -1027,11 +1027,11 @@ void redumper_protection_cd(Options &options)
 
 	std::fstream scm_fs(scra_path, std::fstream::in | std::fstream::binary);
 	if(!scm_fs.is_open())
-		throw_line(fmt::format("unable to open file ({})", scra_path.filename().string()));
+		throw_line(std::format("unable to open file ({})", scra_path.filename().string()));
 
 	std::fstream state_fs(state_path, std::fstream::in | std::fstream::binary);
 	if(!state_fs.is_open())
-		throw_line(fmt::format("unable to open file ({})", state_path.filename().string()));
+		throw_line(std::format("unable to open file ({})", state_path.filename().string()));
 
 	std::string protection("N/A");
 
@@ -1122,7 +1122,7 @@ void redumper_protection_cd(Options &options)
 
 						if(range.second > range.first)
 						{
-							protection = fmt::format("PS2/Datel {}, C2: {}, range: {}-{}", protected_filename, range.second - range.first, range.first, range.second - 1);
+							protection = std::format("PS2/Datel {}, C2: {}, range: {}-{}", protected_filename, range.second - range.first, range.first, range.second - 1);
 
 							auto skip_ranges = string_to_ranges(options.skip);
 							skip_ranges.push_back(range);
@@ -1161,11 +1161,11 @@ void redumper_split_cd(const Options &options)
 
 	std::fstream scm_fs(scra_path, std::fstream::in | std::fstream::binary);
 	if(!scm_fs.is_open())
-		throw_line(fmt::format("unable to open file ({})", scra_path.filename().string()));
+		throw_line(std::format("unable to open file ({})", scra_path.filename().string()));
 
 	std::fstream state_fs(state_path, std::fstream::in | std::fstream::binary);
 	if(!state_fs.is_open())
-		throw_line(fmt::format("unable to open file ({})", state_path.filename().string()));
+		throw_line(std::format("unable to open file ({})", state_path.filename().string()));
 
 	// TOC
 	std::vector<uint8_t> toc_buffer = read_vector(toc_path);
@@ -1190,7 +1190,7 @@ void redumper_split_cd(const Options &options)
 	{
 		std::fstream fs(sub_path, std::fstream::in | std::fstream::binary);
 		if(!fs.is_open())
-			throw_line(fmt::format("unable to open file ({})", sub_path.filename().string()));
+			throw_line(std::format("unable to open file ({})", sub_path.filename().string()));
 
 		subq.resize(sectors_count);
 		std::vector<uint8_t> sub_buffer(CD_SUBCODE_SIZE);
@@ -1616,7 +1616,7 @@ void redumper_split_cd(const Options &options)
 
 	// check tracks
 	if(!check_tracks(toc, scm_fs, state_fs, offset_manager, skip_ranges, scrap, options) && !options.force_split)
-		throw_line(fmt::format("data errors detected, unable to continue"));
+		throw_line(std::format("data errors detected, unable to continue"));
 
 	// write tracks
 	std::vector<TrackEntry> track_entries;
@@ -1630,30 +1630,30 @@ void redumper_split_cd(const Options &options)
 		cue_sheets.resize(toc.cd_text_lang.size());
 		for(uint32_t i = 0; i < toc.cd_text_lang.size(); ++i)
 		{
-			cue_sheets[i] = i ? fmt::format("{}_{:02X}.cue", options.image_name, toc.cd_text_lang[i]) : fmt::format("{}.cue", options.image_name);
+			cue_sheets[i] = i ? std::format("{}_{:02X}.cue", options.image_name, toc.cd_text_lang[i]) : std::format("{}.cue", options.image_name);
 			LOG_F("{}... ", cue_sheets[i]);
 
 			if(std::filesystem::exists(std::filesystem::path(options.image_path) / cue_sheets[i]) && !options.overwrite)
-				throw_line(fmt::format("file already exists ({})", cue_sheets[i]));
+				throw_line(std::format("file already exists ({})", cue_sheets[i]));
 
 			std::fstream fs(std::filesystem::path(options.image_path) / cue_sheets[i], std::fstream::out);
 			if(!fs.is_open())
-				throw_line(fmt::format("unable to create file ({})", cue_sheets[i]));
+				throw_line(std::format("unable to create file ({})", cue_sheets[i]));
 			toc.PrintCUE(fs, options.image_name, i);
 			LOG("done");
 		}
 	}
 	else
 	{
-		cue_sheets.push_back(fmt::format("{}.cue", options.image_name));
+		cue_sheets.push_back(std::format("{}.cue", options.image_name));
 		LOG_F("{}... ", cue_sheets.front());
 
 		if(std::filesystem::exists(std::filesystem::path(options.image_path) / cue_sheets.front()) && !options.overwrite)
-			throw_line(fmt::format("file already exists ({})", cue_sheets.front()));
+			throw_line(std::format("file already exists ({})", cue_sheets.front()));
 
 		std::fstream fs(std::filesystem::path(options.image_path) / cue_sheets.front(), std::fstream::out);
 		if(!fs.is_open())
-			throw_line(fmt::format("unable to create file ({})", cue_sheets.front()));
+			throw_line(std::format("unable to create file ({})", cue_sheets.front()));
 		toc.PrintCUE(fs, options.image_name);
 		LOG("done");
 		LOG("");
@@ -1683,7 +1683,7 @@ void redumper_split_cd(const Options &options)
 		std::filesystem::path cue_path(std::filesystem::path(options.image_path) / c);
 		std::fstream ifs(cue_path, std::fstream::in);
 		if(!ifs.is_open())
-			throw_line(fmt::format("unable to open file ({})", cue_path.filename().string()));
+			throw_line(std::format("unable to open file ({})", cue_path.filename().string()));
 		std::string line;
 		while(std::getline(ifs, line))
 			LOG("{}", line);
@@ -1698,7 +1698,7 @@ std::list<std::pair<std::string, bool>> cue_get_entries(const std::filesystem::p
 
 	std::fstream fs(cue_path, std::fstream::in);
 	if(!fs.is_open())
-		throw_line(fmt::format("unable to open file ({})", cue_path.filename().string()));
+		throw_line(std::format("unable to open file ({})", cue_path.filename().string()));
 
 	std::pair<std::string, bool> entry;
 	std::string line;
